remove auto frame rate for video capture Fixes #25

+++ modules/olvis3/projects/OlvisExecGuiPlugin/src/VideoDisplay/VideoRecorder.h
@@ -57,7 +57,7 @@ protected:
     virtual void run();
 
     int mRate; // rate in ms (<=0: auto)
-    double mFps;
+    int mFps;
 
     QRect mRect;
     VideoDisplayWidget* mWidget;
